+++ CommercialApplication/DomainLayer/Services/OrderServices/IOrderService.cs
@@ -10,7 +10,7 @@ namespace CommercialApplicationCommand.DomainLayer.Services.OrderServices
         Order SelectById(IDbConnection connection, long id, IDbTransaction transaction = null);
         IEnumerable<Order> SelectByDay(IDbConnection connection, string day, IDbTransaction transaction = null);
         void Delete(IDbConnection connection, long id, IDbTransaction transaction = null);
-        long Insert(IDbConnection connection, Order order, IDbTransaction transaction = null);
+        int Insert(IDbConnection connection, Order order, IDbTransaction transaction = null);
         void Update(IDbConnection connection, Order order, IDbTransaction transaction = null);
         bool Exists(IDbConnection connection, long id, IDbTransaction transaction = null);        
         long SelectOrderIdWithMaxSumValueByDay(IDbConnection connection, IEnumerable<Order> orders, IDbTransaction transaction = null);
